pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mimic_viewer"
version = "0.1.0"
description = "A package for viewing Robot data using Rerun"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "German Rodriguez", email = "german.rodriguez@mimicrobotics.com"},
]
keywords = ["rerun", "logging", "embodiment", "visualization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "zarr==2.18",
    "rerun-sdk",
    "scipy",
    "fsspec",
    "gcsfs",
    "rerun-loader-urdf @ git+https://github.com/rerun-io/rerun-loader-python-example-urdf.git",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
]

web = [
    "fastapi",
    "uvicorn[standard]",
    "requests",
    "cloud-sql-python-connector[pg8000]",
    "python-dotenv",
]

[project.urls]
Homepage = "https://mimicrobotics.com"
Repository = "https://github.com/mimicrobotics/mimic_viewer"

[tool.hatch.build.targets.wheel]
packages = ["src/mimic_viewer"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 

[tool.hatch.metadata]
allow-direct-references = true 